Move make_task to upper level + return schedule id
